Add Atomic-OS Version
Every ROM claims it as their own and in turn making it very difficult to
figure out who the real author is.

It might of originated with CM so credit to them if so but I'm not entirely sure
who exactly within the CM org.

If someone knows, hit me up and we'll give them a gold star and a hand shake!

Change-Id: I8007284251d8e8cf81ac43a3da61617809d70d67
Signed-off-by: AKabhishek <abhishekkaushik.star@gmail.com>